+++ robot.cpp
@@ -1,6 +1,6 @@
 #include "robot.h"
 
-robot::robot(int _id, double x, double y) {
+robot::robot(int _id, float x, float y) {
     this->ID = _id;
     this->positionX = x;
     this->positionY = y;
