+++ station.cpp
@@ -1,6 +1,6 @@
 #include "station.h"
 
-station::station(int _type, int _id, double x, double y):
+station::station(int _type, int _id, float x, float y):
         type(_type), id(_id), positionX(x), positionY(y) {}
 station::station()=default;
