+++ station.h
@@ -14,7 +14,7 @@ using namespace std;
 struct Ask{
     int RID{-1};
     int timeStamp{-1};
-    double value{-1};
+    float value{-1};
 };
 
 class station {
@@ -22,8 +22,8 @@ public:
 
     const int id{-1};
     const int type{0};
-    const double positionX{0.0};
-    const double positionY{0.0};
+    const float positionX{0.0};
+    const float positionY{0.0};
 
     int timeRemain{-1};
     int proState{0};
@@ -35,7 +35,7 @@ public:
     // 各个端口将接收到对应机器人买卖请求
     array<vector<Ask>, 8> portRecvAskTime;
 
-    station(int, int, double, double);
+    station(int, int, float, float);
     station();
 
     string to_string();
